+++ csfunctions/events/engineering_change_released.py
@@ -7,7 +7,7 @@ from csfunctions.objects import Document, EngineeringChange, Part
 from .base import BaseEvent, EventNames
 
 
-class EngineeringChangeReleaseData(BaseModel):
+class EngineeringChangeReleasedData(BaseModel):
     documents: list[Document] = Field(..., description="List of included documents.")
     parts: list[Part] = Field(..., description="List of included parts.")
     engineering_changes: list[EngineeringChange] = Field(
@@ -15,6 +15,6 @@ class EngineeringChangeReleaseData(BaseModel):
     )
 
 
-class EngineeringChangeRelease(BaseEvent):
-    name: Literal[EventNames.ENGINEERING_CHANGE_RELEASE] = EventNames.ENGINEERING_CHANGE_RELEASE
-    data: EngineeringChangeReleaseData
+class EngineeringChangeReleasedEvent(BaseEvent):
+    name: Literal[EventNames.ENGINEERING_CHANGE_RELEASED] = EventNames.ENGINEERING_CHANGE_RELEASED
+    data: EngineeringChangeReleasedData