+++ csfunctions/events/part_release_check.py
@@ -5,13 +5,13 @@ from pydantic import BaseModel, Field
 from csfunctions.objects import Document, Part
 
 from .base import BaseEvent, EventNames
-from .dialog_data import PartReleaseDialogData
+from .dialog_data import PartReleasedDialogData
 
 
 class PartReleaseCheckData(BaseModel):
     parts: list[Part] = Field(..., description="List of parts that will be released.")
-    attached_documents: list[Document] = Field(..., description="List of documents that are referenced by the parts.")
-    dialog_data: PartReleaseDialogData
+    documents: list[Document] = Field(..., description="List of documents that are referenced by the parts.")
+    dialog_data: PartReleasedDialogData
 
 
 class PartReleaseCheckEvent(BaseEvent):